+++ manual/release-notes.rst
@@ -6,7 +6,7 @@ Release Notes
 For a detailed list of changes, please see the file
 :file:`ChangeLog` in the source distribution.
 
-10.6.0: February 8, 2022
+10.6.0: February 9, 2022
   - Preparation for replacement of ``PointerHolder``
 
     The next major release of qpdf will replace ``PointerHolder`` with